Refactor test code so it can be excluded from code coverage
