+++ .github/workflows/ci.yml
@@ -36,7 +36,7 @@ jobs:
       - name: Generate code coverage
         env:
           RUST_BACKTRACE: "1"
-        run: cargo llvm-cov --all-features --lcov --output-path lcov.info
+        run: cargo llvm-cov --all-features --lcov --ignore-filename-regex tests --output-path lcov.info
 
       - name: Upload code coverage results
         uses: codecov/codecov-action@v3
