+++ src/tests/fixtures/Purple Square.psd
