+++ src/tests/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2020 Adobe. All rights reserved.
+// Copyright 2022 Adobe. All rights reserved.
 // This file is licensed to you under the Apache License,
 // Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0)
 // or the MIT license (http://opensource.org/licenses/MIT),
@@ -11,20 +11,10 @@
 // specific language governing permissions and limitations under
 // each license.
 
-#![deny(warnings)]
-#![deny(missing_docs)]
-#![doc = include_str!("../README.md")]
+// Tests are grouped under this module so as to avoid
+// having the test code itself included in coverage numbers.
 
-mod ffi;
-
-#[cfg(test)]
-mod tests;
-
+mod fixtures;
 mod xmp_date_time;
 mod xmp_file;
 mod xmp_meta;
-pub mod xmp_ns;
-
-pub use xmp_date_time::XmpDateTime;
-pub use xmp_file::{OpenFileOptions, XmpFile, XmpFileError};
-pub use xmp_meta::XmpMeta;
