+++ src/xmp_file.rs
@@ -282,110 +282,3 @@ fn path_to_cstr(path: &Path) -> Option<CString> {
         None => None,
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use std::{
-        env, fs,
-        path::{Path, PathBuf},
-    };
-
-    use tempfile::tempdir;
-
-    use crate::{xmp_ns, XmpDateTime};
-
-    fn fixture_path(name: &str) -> String {
-        let root_dir = &env::var("CARGO_MANIFEST_DIR").unwrap();
-
-        let mut path = PathBuf::from(root_dir);
-        path.push("tests/fixtures");
-        path.push(name);
-        path.to_str().unwrap().to_string()
-    }
-
-    fn temp_copy_of_fixture(tempdir: &Path, name: &str) -> String {
-        let fixture_src = fixture_path(name);
-        let fixture_path = Path::join(tempdir, name);
-        let fixture_copy = fixture_path.as_path();
-
-        fs::copy(fixture_src, fixture_copy).unwrap();
-        fixture_copy.display().to_string()
-    }
-
-    #[test]
-    fn open_and_edit_file() {
-        let tempdir = tempdir().unwrap();
-        let purple_square = temp_copy_of_fixture(tempdir.path(), "Purple Square.psd");
-
-        {
-            let mut f = XmpFile::new();
-
-            assert!(f
-                .open_file(
-                    &purple_square,
-                    OpenFileOptions::default().for_update().use_smart_handler()
-                )
-                .is_ok());
-
-            let opt_m = f.xmp();
-            assert!(opt_m.is_some());
-
-            XmpMeta::register_namespace("http://purl.org/dc/terms/", "dcterms");
-
-            let mut m = opt_m.unwrap();
-            m.set_property("http://purl.org/dc/terms/", "provenance", "blah");
-
-            assert!(m.does_property_exist("http://purl.org/dc/terms/", "provenance"));
-            assert!(!m.does_property_exist("http://purl.org/dc/terms/", "provenancx"));
-
-            if m.does_property_exist(xmp_ns::XMP, "MetadataDate") {
-                let updated_time = XmpDateTime::current();
-                m.set_property_date(xmp_ns::XMP, "MetadataDate", &updated_time);
-            }
-
-            assert!(f.can_put_xmp(&m));
-            f.put_xmp(&m);
-
-            f.close();
-        }
-
-        // Let's make sure we actually wrote to the file.
-        {
-            let mut f = XmpFile::new();
-
-            assert!(f
-                .open_file(
-                    &purple_square,
-                    OpenFileOptions::default().for_update().use_smart_handler()
-                )
-                .is_ok());
-
-            let m = f.xmp().unwrap();
-
-            assert_eq!(
-                m.property("http://purl.org/dc/terms/", "provenance")
-                    .unwrap(),
-                "blah"
-            );
-            assert_eq!(m.property("http://purl.org/dc/terms/", "provenancx"), None);
-        }
-    }
-
-    #[test]
-    fn open_fail() {
-        let bad_path = PathBuf::from("doesnotexist.jpg");
-
-        {
-            let mut f = XmpFile::new();
-
-            assert!(f
-                .open_file(
-                    &bad_path,
-                    OpenFileOptions::default().for_update().use_smart_handler()
-                )
-                .is_err());
-        }
-    }
-}
