+++ src/xmp_meta.rs
@@ -201,52 +201,3 @@ impl XmpMeta {
         r != 0
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use std::{env, path::PathBuf};
-
-    fn fixture_path(name: &str) -> PathBuf {
-        let root_dir = &env::var("CARGO_MANIFEST_DIR").unwrap();
-        let mut path = PathBuf::from(root_dir);
-        path.push("tests/fixtures");
-        path.push(name);
-        path
-    }
-
-    #[test]
-    fn new_empty() {
-        let mut _m = XmpMeta::new();
-    }
-
-    #[test]
-    fn from_file() {
-        let m = XmpMeta::from_file(fixture_path("Purple Square.psd")).unwrap();
-
-        assert_eq!(
-            m.property("http://ns.adobe.com/xap/1.0/", "CreatorTool")
-                .unwrap(),
-            "Adobe Photoshop CS2 Windows"
-        );
-
-        assert_eq!(
-            m.property("http://ns.adobe.com/photoshop/1.0/", "ICCProfile")
-                .unwrap(),
-            "Dell 1905FP Color Profile"
-        );
-
-        assert!(m
-            .property("http://ns.adobe.com/photoshop/1.0/", "ICCProfilx")
-            .is_none());
-    }
-
-    #[test]
-    fn register_namespace() {
-        assert_eq!(
-            XmpMeta::register_namespace("http://purl.org/dc/terms/", "dcterms"),
-            "dcterms:"
-        );
-    }
-}
